refactor: some clean-up
